pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "delfos-nl2sql"
version = "0.1.0"
description = "Natural Language to SQL pipeline with multi-step orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "azure-identity>=1.15.0",
    "anthropic>=0.18.0",
    "mcp>=0.9.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "agent-framework>=0.1.0",
    "agent-framework-azure-ai>=0.1.0",
    "pandas>=3.0.0rc0",
    "ipykernel>=7.1.0",
    "pyodbc>=5.3.0",
    "agent-framework-anthropic>=1.0.0b251204",
]

[tool.uv]
prerelease = "allow"

[dependency-groups]
dev = [
    "ruff>=0.14.8",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.19.0",
    "bandit>=1.9.2",
    "types-python-dateutil>=2.9.0.20251115",
    "types-requests>=2.32.4.20250913",
]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "B", "I", "UP", "SIM"]
lint.ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

strict = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true

files = ["src"]
exclude = "(^src/examples/)"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.bandit]
targets = ["src"]
skips = ["B101"]
